deduplicated Voxels partition and VoxelLoader
